pyproject.toml
[tool.poetry]
name = "wiki_documental"
version = "1.0.0"
description = "Sistema automatizado de ingesta documental para generación de wiki navegable con Docsify"
authors = ["Jose Manuel Lago <jose.lago.varela@gmail.com>"]
license = "MIT"
readme = "README.md"

[[tool.poetry.packages]]
include = "wiki"
from = "src"

[tool.poetry.dependencies]
python = "^3.11"
python-docx = "^1.1"
PyYAML = "^6.0"
rich = "^13.7"
python-slugify = "^8.0"
tqdm = "^4.66"
typer = {version = "^0.9", extras = ["all"]}
"pdfminer.six" = {version = "^20221105", optional = true}
pdf2image = {version = "^1.17", optional = true}
pytesseract = {version = "^0.3", optional = true}

[tool.poetry.extras]
pdf = ["pdfminer.six", "pdf2image", "pytesseract"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
types-python-slugify = "^8.0"

[tool.poetry.scripts]
wiki = "wiki.cli:app"
